solo_project/youtube_trend_ai_analyzer/requirements.txt
google-genai
supabase
pandas
python-dotenv
